refactor: transforma os componentes em pastas
