+++ src/pages/404.js
@@ -6,8 +6,8 @@ import Layout from '../layouts/Layout';
 const NotFoundPage = () => (
   <Layout>
     <SEO title="404: Not found" />
-    <h1>NOT FOUND</h1>
-    <p>You just hit a route that doesn&#39;t exist... the sadness.</p>
+    <h1>Oops</h1>
+    <p>You just hit a route that doesn&#39;t exist... </p>
   </Layout>
 );
